refactor: hoist 'agents.find_skill_toolset'
  ... from 'views.agui'
